+++ packages/default-layout/src/types/StoreProps.ts
@@ -7,5 +7,6 @@
  */
 
 export interface StoreProps {
+    isCurrentTabOpened: boolean;
     currentTab: number;
 }